requirements.txt
requests
python-dotenv
torch
torchvision
pillow
scikit-learn
matplotlib
pyyaml
